Extract main into separate file
