+++ core/perf_test/Benchmark_ViewCopy.cpp
@@ -1,12 +1,56 @@
+/*
+//@HEADER
+// ************************************************************************
+//
+//                        Kokkos v. 3.0
+//       Copyright (2020) National Technology & Engineering
+//               Solutions of Sandia, LLC (NTESS).
+//
+// Under the terms of Contract DE-NA0003525 with NTESS,
+// the U.S. Government retains certain rights in this software.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are
+// met:
+//
+// 1. Redistributions of source code must retain the above copyright
+// notice, this list of conditions and the following disclaimer.
+//
+// 2. Redistributions in binary form must reproduce the above copyright
+// notice, this list of conditions and the following disclaimer in the
+// documentation and/or other materials provided with the distribution.
+//
+// 3. Neither the name of the Corporation nor the names of the
+// contributors may be used to endorse or promote products derived from
+// this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY NTESS "AS IS" AND ANY
+// EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+// PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL NTESS OR THE
+// CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+// EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
+// Questions? Contact Christian R. Trott (crtrott@sandia.gov)
+//
+// ************************************************************************
+//@HEADER
+*/
+
 #include <benchmark/benchmark.h>
+
+#include <Kokkos_Core.hpp>
+#include <PerfTest_ViewCopy.hpp>  // Test::deepcopy_view
 
 #include <cmath>
 
-#include <Kokkos_Core.hpp>
-
-#include <PerfTest_ViewCopy.hpp>
-
 // replicates core/perf_test/PerfTest_ViewCopy_d8.cpp
+namespace Test {
 
 void report_results(benchmark::State& state, double time) {
   state.SetIterationTime(time);
@@ -74,26 +118,4 @@ BENCHMARK(ViewDeepCopy_Rank8_Raw<Kokkos::LayoutRight, Kokkos::LayoutLeft>)
     ->UseManualTime();
 #endif
 
-std::string custom_context() {
-  std::ostringstream msg;
-  Kokkos::print_configuration(msg);
-  return msg.str();
-}
-
-// Run benchmarks
-// BENCHMARK_MAIN() + Kokkos init / finalize
-int main(int argc, char** argv) {
-  benchmark::Initialize(&argc, argv);
-  if (benchmark::ReportUnrecognizedArguments(argc, argv)) return 1;
-  Kokkos::initialize(argc, argv);
-  benchmark::SetDefaultTimeUnit(benchmark::kSecond);
-  benchmark::AddCustomContext("Kokkos configuration", custom_context());
-
-  benchmark::RunSpecifiedBenchmarks();
-  // REMOVE_ME: Run the vanilla test for comparison
-  Test::run_deepcopyview_tests8<Kokkos::LayoutRight, Kokkos::LayoutLeft>(10, 1);
-
-  benchmark::Shutdown();
-  Kokkos::finalize();
-  return 0;
-}
+}  // namespace Test
